itest: move channel backup tests to own file
